Add Vehicles button component
